+++ changelog.md
@@ -6,6 +6,7 @@ and this project adheres to http://semver.org/spec/v2.0.0.html.
 
 ## [unreleased]
 
+- Add option --tab-width
 - Reset terminal coloring if trimmed
 
 ## [0.1.0] 2022-11-30